requirements.txt
requests==2.31.0
flask==3.0.3
google-generativeai==0.7.2
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.2
